# EGAD-Openapi-Docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /rel/{index}/select/table.{table}/columns.{columns}:
    get:
      tags:
      - relational-rest-index
      summary: "from database at index {index}, select from table table.{table} the\
        \ value in multiple columns {columns}(comma separated)"
      operationId: getMultiple
      parameters:
      - name: index
        in: path
        description: the index of the database to make a request to
        required: true
        schema:
          type: integer
          format: int32
      - name: table
        in: path
        description: sql table name
        required: true
        schema:
          type: string
      - name: columns
        in: path
        description: array of columns to get
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: "table, value, and row exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rel/select/table.{table}/columns.{columns}:
    get:
      tags:
      - relational-rest
      summary: "select from table table.{table} the value in multiple columns {columns}(comma\
        \ separated)"
      operationId: getMultiple_1
      parameters:
      - name: table
        in: path
        description: sql table name
        required: true
        schema:
          type: string
      - name: columns
        in: path
        description: array of columns to get
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: "table, value, and row exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTable'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /kv/{index}/type/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "get the type of key {key} from database in index {index}"
      operationId: getType
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringData'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/setex/key.{key}/value.{value}/t.{time}:
    get:
      tags:
      - kv-rest-index
      summary: "Set key {key} to hold the string value {value} and set key to timeout\
        \ after a given number of seconds {t}"
      operationId: setEx
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: value
        in: path
        required: true
        schema:
          type: string
      - name: time
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/set/key.{key}/value.{value}:
    get:
      tags:
      - kv-rest-index
      summary: "Set key {key} to have value {value}"
      operationId: setKey
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: value
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/incrby/key.{key}/incr.{incr}:
    get:
      tags:
      - kv-rest-index
      summary: "in the key value database in the index {index}, increment key {key}\
        \ by increment amount {inc}"
      operationId: incr
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: incr
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LongData'
  /kv/{index}/incr/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "in the key value database in the index {index}, increment key {key}\
        \ by 1"
      operationId: incr_1
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LongData'
  /kv/{index}/getMul/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "get multiple keys {key} (comma separated list) from database in index\
        \ {index}"
      operationId: getMultiple_2
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/get/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "get key {key}"
      operationId: getKey_2
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringData'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/expire/key.{key}/t.{time}:
    get:
      tags:
      - kv-rest-index
      summary: "Set key {key} to timeout after a given number of seconds {t}"
      operationId: expire
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      - name: time
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/exists/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "in the key value database in the index {index}, increment key {key}\
        \ by 1"
      operationId: exists
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoolData'
  /kv/{index}/del/key.{key}:
    get:
      tags:
      - kv-rest-index
      summary: "delete key {key} in database with index {index}"
      operationId: del
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/{index}/addMul/key.{key}/value.{value}:
    get:
      tags:
      - kv-rest-index
      summary: "add multiple keys {key} (comma separated list) with values {value}\
        \ (comma separated list) to database in position {index}"
      operationId: addMul
      parameters:
      - name: index
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      - name: value
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kv/type/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "get the type of key {key} from all databases"
      operationId: getType_1
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListString'
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /kv/getMul/key.{key}:
    get:
      tags:
      - kv-rest
      summary: "get multiple keys {key} (comma separated list) from all databases"
      operationId: getMultiple_3
      parameters:
      - name: key
        in: path
        required: true
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTable'
        "304":
          description: nothing changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        rc:
          type: string
          enum:
          - OK
          - CREATED
          - ACCEPTED
          - NO_CONTENT
          - NOT_MODIFIED
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - TOO_MANY_REQUESTS
          - SERVICE_UNAVAILABLE
        responseCode:
          type: string
          enum:
          - OK
          - CREATED
          - ACCEPTED
          - NO_CONTENT
          - NOT_MODIFIED
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - TOO_MANY_REQUESTS
          - SERVICE_UNAVAILABLE
    Table:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
    ListTable:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
    ListString:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
    StringData:
      type: object
      properties:
        data:
          type: string
...
    LongData:
      type: object
      properties:
        data:
          type: integer
          format: int64
    BoolData:
      type: object
      properties:
        data:
          type: boolean
...